Move drawing functions from the visualizer to the gui folder so they can be shared
