+++ src/software/gui/drawing/BUILD
@@ -1,4 +1,4 @@
-package(default_visibility = ["//visibility:public"])
+package(default_visibility = ["//software/gui:__subpackages__"])
 
 cc_library(
     name = "colors",
