+++ src/software/gui/drawing/colors.h
