+++ src/software/gui/drawing/draw_functions.h
