+++ src/software/gui/drawing/field.cpp
@@ -1,4 +1,4 @@
-#include "software/gui/visualizer/drawing/field.h"
+#include "software/gui/drawing/field.h"
 
 #include "shared/constants.h"
