+++ src/software/gui/drawing/geom.h
