+++ src/software/gui/drawing/obstacle_artist.h
@@ -4,9 +4,9 @@
 
 #include "software/ai/navigator/obstacle/obstacle.h"
 #include "software/ai/navigator/obstacle/obstacle_visitor.h"
-#include "software/gui/visualizer/drawing/colors.h"
-#include "software/gui/visualizer/drawing/draw_functions.h"
-#include "software/gui/visualizer/drawing/geom.h"
+#include "software/gui/drawing/colors.h"
+#include "software/gui/drawing/draw_functions.h"
+#include "software/gui/drawing/geom.h"
 #include "software/gui/geometry_conversion/geometry_conversion.h"
 
 /**
