+++ src/software/gui/drawing/robot.h
@@ -2,7 +2,7 @@
 
 #include <QtWidgets/QGraphicsScene>
 
-#include "software/gui/visualizer/drawing/geom.h"
+#include "software/gui/drawing/geom.h"
 #include "software/gui/geometry_conversion/geometry_conversion.h"
 #include "software/new_geom/segment.h"
 #include "software/world/robot.h"
