+++ src/software/gui/drawing/team.cpp
@@ -1,7 +1,7 @@
-#include "software/gui/visualizer/drawing/team.h"
+#include "software/gui/drawing/team.h"
 
-#include "software/gui/visualizer/drawing/colors.h"
-#include "software/gui/visualizer/drawing/robot.h"
+#include "software/gui/drawing/colors.h"
+#include "software/gui/drawing/robot.h"
 
 void drawTeam(QGraphicsScene* scene, const Team& team, QColor color)
 {
