+++ src/software/gui/visualizer/BUILD
@@ -10,8 +10,8 @@ cc_library(
         "//software/ai/hl/stp:play_info",
         "//software/backend:robot_status",
         "//software/geom",
-        "//software/gui/visualizer/drawing:draw_functions",
-        "//software/gui/visualizer/drawing:world",
+        "//software/gui/drawing:draw_functions",
+        "//software/gui/drawing:world",
         "//software/gui/visualizer/widgets:visualizer",
         "//software/multithreading:thread_safe_buffer",
         "//software/multithreading:threaded_observer",
