+++ src/software/gui/visualizer/widgets/BUILD
@@ -12,10 +12,10 @@ qt_cc_library(
         "//software/ai/hl/stp:play_info",
         "//software/backend:robot_status",
         "//software/geom",
+        "//software/gui/drawing:draw_functions",
         "//software/gui/generic_widgets/robot_status",
         "//software/gui/generic_widgets/world_view",
         "//software/gui/geometry_conversion",
-        "//software/gui/visualizer/drawing:draw_functions",
         "//software/gui/visualizer/ui:main_widget",
         "//software/logger",
         "//software/proto:sensor_msg_cc_proto",
